// mobile/api/gen/src/models/MessagePublicDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Offlinery
 * API of Offlinery
 *
 * The version of the OpenAPI document: 1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface MessagePublicDTO
 */
export interface MessagePublicDTO {
    /**
     * Message id.
     * @type {string}
     * @memberof MessagePublicDTO
     */
    id: string;
    /**
     * Message sent by user.
     * @type {string}
     * @memberof MessagePublicDTO
     */
    content: string;
    /**
     * Who sent the message?
     * @type {string}
     * @memberof MessagePublicDTO
     */
    senderUserId: string;
    /**
     * The date time the message was sent.
     * @type {string}
     * @memberof MessagePublicDTO
     */
    sentAt: string;
}

/**
 * Check if a given object implements the MessagePublicDTO interface.
 */
export function instanceOfMessagePublicDTO(
    value: object,
): value is MessagePublicDTO {
    if (!("id" in value) || value["id"] === undefined) return false;
    if (!("content" in value) || value["content"] === undefined) return false;
    if (!("senderUserId" in value) || value["senderUserId"] === undefined)
        return false;
    if (!("sentAt" in value) || value["sentAt"] === undefined) return false;
    return true;
}

export function MessagePublicDTOFromJSON(json: any): MessagePublicDTO {
    return MessagePublicDTOFromJSONTyped(json, false);
}

export function MessagePublicDTOFromJSONTyped(
    json: any,
    ignoreDiscriminator: boolean,
): MessagePublicDTO {
    if (json == null) {
        return json;
    }
    return {
        id: json["id"],
        content: json["content"],
        senderUserId: json["senderUserId"],
        sentAt: json["sentAt"],
    };
}

export function MessagePublicDTOToJSON(value?: MessagePublicDTO | null): any {
    if (value == null) {
        return value;
    }
    return {
        id: value["id"],
        content: value["content"],
        senderUserId: value["senderUserId"],
        sentAt: value["sentAt"],
    };
}