// mobile/api/gen/src/models/UpdateEncounterStatusDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Offlinery
 * API of Offlinery
 *
 * The version of the OpenAPI document: 1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface UpdateEncounterStatusDTO
 */
export interface UpdateEncounterStatusDTO {
    /**
     * New encounter status
     * @type {string}
     * @memberof UpdateEncounterStatusDTO
     */
    status: UpdateEncounterStatusDTOStatusEnum;
    /**
     * Encounter ID
     * @type {string}
     * @memberof UpdateEncounterStatusDTO
     */
    encounterId: string;
}

/**
 * @export
 */
export const UpdateEncounterStatusDTOStatusEnum = {
    not_met: "not_met",
    met_not_interested: "met_not_interested",
    met_interested: "met_interested",
} as const;
export type UpdateEncounterStatusDTOStatusEnum =
    (typeof UpdateEncounterStatusDTOStatusEnum)[keyof typeof UpdateEncounterStatusDTOStatusEnum];

/**
 * Check if a given object implements the UpdateEncounterStatusDTO interface.
 */
export function instanceOfUpdateEncounterStatusDTO(
    value: object,
): value is UpdateEncounterStatusDTO {
    if (!("status" in value) || value["status"] === undefined) return false;
    if (!("encounterId" in value) || value["encounterId"] === undefined)
        return false;
    return true;
}

export function UpdateEncounterStatusDTOFromJSON(
    json: any,
): UpdateEncounterStatusDTO {
    return UpdateEncounterStatusDTOFromJSONTyped(json, false);
}

export function UpdateEncounterStatusDTOFromJSONTyped(
    json: any,
    ignoreDiscriminator: boolean,
): UpdateEncounterStatusDTO {
    if (json == null) {
        return json;
    }
    return {
        status: json["status"],
        encounterId: json["encounterId"],
    };
}

export function UpdateEncounterStatusDTOToJSON(
    value?: UpdateEncounterStatusDTO | null,
): any {
    if (value == null) {
        return value;
    }
    return {
        status: value["status"],
        encounterId: value["encounterId"],
    };
}